+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## 1.2.1-dev
 
+* Add `sequentialAsyncExpand` to sequentially emit events emitted by multiple
+  sub streams created by a callback without previous sub streams
+  blocking the source stream and without interleaving events from previous sub streams.
+
 ## 1.2.0
 
 -  Add support for emitting the "leading" event in `debounce`.